pyproject.toml
[project]
name = "snapcheck"
version = "0.0.1"
authors = [
    {name = "BrainVISA team", email = "contact@brainvisa.info"},
]
description = "Control the quality of analysis performed on many neuroimaging data"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "CeCILL-B"}
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved",
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Natural Language :: English",
]
dependencies = [
    "populse-db >= 3",
    "soma-base >= 6",
]

[project.optional-dependencies]
doc = [
    "sphinx >=1.0",
]

[project.urls]
homepage = "http://brainvisa.github.io/snapcheck/"
repository = "https://github.com/brainvisa/snapcheck"
